copilot_web/requirements.txt
flask>=3.0.0
openai>=1.0.0
gunicorn>=21.0.0
playwright>=1.44.0
mpxj
pandas>=2.2.0
python-docx>=1.1.0
pdfplumber>=0.11.0
